# specs/player-public.openapi.yaml
openapi: 3.1.1
info:
  title: SpaceY Player and Public API
  version: 1.0.0
  description: Server-authoritative player API and read-only partner API. Gameplay completion is never accepted from clients.
  license: { name: Proprietary, identifier: LicenseRef-SpaceY-Proprietary }
servers:
  - url: https://spacey.aima.space
...
paths:
...
  /api/v1/bootstrap:
    get:
      tags: [Player]
      operationId: getPlayerBootstrap
      summary: Read current server-owned player state
      security: [{ playerAccess: [] }]
      responses:
        "200":
          description: Current server-owned profile, economy, content and build state.
          content:
            application/json: { schema: { $ref: "#/components/schemas/Bootstrap" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/v1/builds/{buildId}/commands:
    put:
      tags: [Player]
      operationId: applyShipBuildCommands
      summary: Apply commands as a new immutable build revision
      security: [{ playerAccess: [] }]
      parameters:
        - $ref: "#/components/parameters/BuildId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BuildCommands" }
      responses:
        "200":
          description: New immutable build revision.
          content:
            application/json: { schema: { $ref: "#/components/schemas/ShipBuild" } }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/builds/legacy-import-proposals:
    post:
      tags: [Player]
      operationId: importLegacyBuildV3
      summary: Import one legacy schema-v3 build
      description: One-time build-only import. Wallet, rewards and mission results are ignored.
      security: [{ playerAccess: [] }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: "#/components/schemas/LegacyBuildImport" } }
      responses:
        "200": { description: Imported or previously imported build. }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/mission-attempts:
    post:
      tags: [Battles]
      operationId: createMissionAttempt
      summary: Create an authoritative PvE attempt
      security: [{ playerAccess: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [missionId, shipBuildRevisionId, idempotencyKey]
              properties:
                missionId: { type: string }
                shipBuildRevisionId: { type: string, format: uuid }
                idempotencyKey: { type: string, minLength: 16, maxLength: 128 }
      responses:
        "201":
          description: Server-owned attempt created. A one-time ticket is issued only by the connection endpoint.
          content:
            application/json: { schema: { $ref: "#/components/schemas/MissionAttemptStatus" } }
        "422": { $ref: "#/components/responses/Conflict" }
  /api/v1/mission-attempts/{attemptId}:
    get:
      tags: [Battles]
      operationId: getMissionAttempt
      summary: Read attempt and reconnect status
      security: [{ playerAccess: [] }]
      parameters:
        - name: attemptId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Authoritative attempt/reconnect/result status.
          content:
            application/json: { schema: { $ref: "#/components/schemas/MissionAttemptStatus" } }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/mission-attempts/{attemptId}/reconnect:
    post:
      tags: [Battles]
      operationId: reconnectMissionAttempt
      summary: Issue a fresh one-time reconnect ticket
      security: [{ playerAccess: [] }]
      parameters:
        - name: attemptId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "201":
          description: Fresh one-time reconnect ticket for an owned resumable attempt.
          content:
            application/json: { schema: { $ref: "#/components/schemas/BattleConnection" } }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/mission-attempts/{attemptId}/connection:
    post:
      tags: [Battles]
      operationId: createMissionAttemptConnection
      summary: Rotate and issue a one-time PvE connection ticket
      security: [{ playerAccess: [] }]
      parameters:
        - name: attemptId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "201":
          description: Fresh one-time ticket; any previous ticket is revoked.
          content:
            application/json: { schema: { $ref: "#/components/schemas/BattleConnection" } }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/mission-attempts/{attemptId}/abandon:
    post:
      tags: [Battles]
      operationId: abandonMissionAttempt
      summary: Abandon an owned unfinished PvE attempt
      security: [{ playerAccess: [] }]
      parameters:
        - name: attemptId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "201":
          description: Attempt is terminal and no longer holds the build.
          content:
            application/json: { schema: { $ref: "#/components/schemas/MissionAttemptStatus" } }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    playerAccess:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: __Host-spacey_refresh
    publicApiKey:
      type: apiKey
      in: header
      name: X-API-Key
    oauthClient:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: /public/v1/oauth/token
          scopes:
            catalog:read: Read published catalog.
            leaderboards:read: Read public leaderboards.
            profiles:read: Read consented public profiles.
            stats:read: Read aggregate statistics.
    webhookSignature:
      type: apiKey
      in: header
      name: X-SpaceY-Signature
  parameters:
    BuildId:
      name: buildId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Unauthorized:
      description: Authentication failed.
      content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
    Conflict:
      description: Revision or idempotency conflict.
      content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
    NotFound:
      description: Resource not found or not owned by the caller.
      content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
  schemas:
...
    ApiError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, correlationId]
          properties:
            code: { type: string }
            message: { type: string }
            correlationId: { type: string }
            details: { type: object, additionalProperties: true }
    PlayerProfile:
      type: object
      required: [id, telegramUserId, displayName, avatarUrl, locale, createdAt]
      properties:
        id: { type: string, format: uuid }
        telegramUserId: { type: string, pattern: "^[0-9]+$" }
        displayName: { type: string }
        avatarUrl: { type: [string, "null"], format: uri }
        locale: { type: string }
        createdAt: { type: string, format: date-time }
...
    Wallet:
      type: object
      additionalProperties: false
      required: [credits, scrap, alloy, dataShards]
      properties:
        credits: { type: integer, minimum: 0 }
        scrap: { type: integer, minimum: 0 }
        alloy: { type: integer, minimum: 0 }
        dataShards: { type: integer, minimum: 0 }
    InventoryItem:
      type: object
      additionalProperties: false
      required: [id, definitionId, contentVersion, rarity, state, durability, category, shape, stats, visualKey, installedBuildRevisionId, createdAt]
      properties:
        id: { type: string, format: uuid }
        definitionId: { type: string, minLength: 1 }
        contentVersion: { type: string, minLength: 1 }
        rarity: { enum: [common, uncommon, superRare] }
        state: { enum: [available, installed, damaged, destroyed] }
        durability: { type: integer, minimum: 0, maximum: 10000 }
        category: { type: string, minLength: 1 }
        shape:
          type: object
          additionalProperties: false
          required: [cells]
          properties:
            cells:
              type: array
              minItems: 1
              items:
                type: array
                minItems: 2
                maxItems: 2
                items: { type: integer }
        stats: { type: object, additionalProperties: true }
        visualKey: { type: string, minLength: 1 }
        installedBuildRevisionId:
          oneOf:
            - { type: string, format: uuid }
            - { type: "null" }
        createdAt: { type: string, format: date-time }
...
    Mission:
      type: object
      required: [id, contentVersion, name, type, risk, durationSeconds, objective, rewardPreview]
      properties:
        id: { type: string }
        contentVersion: { type: string }
        name: { type: string }
        type: { enum: [salvage, escort, mining, intercept, defense] }
        risk: { enum: [green, yellow, red] }
        briefing: { type: string }
        durationSeconds: { type: integer }
        objective: { type: object, additionalProperties: true }
        rewardPreview: { type: object, additionalProperties: true }
    ShipBuild:
      type: object
      required: [id, activeRevision, updatedAt]
      properties:
        id: { type: string, format: uuid }
        activeRevision: { type: object, additionalProperties: true }
        updatedAt: { type: string, format: date-time }
    BuildCommands:
      type: object
      additionalProperties: false
      required: [expectedRevision, idempotencyKey, commands]
      properties:
        expectedRevision: { type: integer, minimum: 1 }
        idempotencyKey: { type: string, minLength: 16, maxLength: 128 }
        commands:
          type: array
          minItems: 1
          maxItems: 128
          items: { type: object, required: [type], properties: { type: { enum: [rename, install, move, remove] } } }
    Bootstrap:
      type: object
      additionalProperties: false
      required: [serverTime, profile, wallet, activeBuild, inventory, contentRelease, missions, activeGameplay, capabilities]
      properties:
        serverTime: { type: string, format: date-time }
        profile: { $ref: "#/components/schemas/PlayerProfile" }
        wallet: { $ref: "#/components/schemas/Wallet" }
        activeBuild: { oneOf: [{ $ref: "#/components/schemas/ShipBuild" }, { type: "null" }] }
        inventory:
          type: array
          items: { $ref: "#/components/schemas/InventoryItem" }
        contentRelease: { type: object, additionalProperties: true }
        missions: { type: array, items: { $ref: "#/components/schemas/Mission" } }
        activeGameplay: { type: array, items: { $ref: "#/components/schemas/ActiveGameplay" } }
        capabilities:
          type: object
          additionalProperties: false
          required: [pvpMatchmaking, repair]
          properties:
            pvpMatchmaking: { type: boolean }
            repair: { type: boolean }
    BattleConnection:
      type: object
      required: [sessionId, attemptId, mode, websocketUrl, ticket, ticketExpiresAt, protocolVersion]
      properties:
        sessionId: { type: string, format: uuid }
        attemptId: { type: string, format: uuid }
        mode: { enum: [pve, pvp] }
        websocketUrl: { type: string, format: uri }
        ticket: { type: string }
        ticketExpiresAt: { type: string, format: date-time }
        protocolVersion: { type: string }
...
    MatchmakingTicket:
      type: object
      additionalProperties: false
      required: [id, queue, region, mmr, status, createdAt, expiresAt, match]
      properties:
        id: { type: string, format: uuid }
        queue: { type: string }
        region: { type: string }
        mmr: { type: integer, minimum: 0 }
        status: { enum: [queued, matched, completed, cancelled, expired, failed] }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        match:
          oneOf:
            - { type: "null" }
            - type: object
              additionalProperties: false
              required: [matchId, sessionId, attemptId, runtimeState, connection]
              properties:
                matchId: { type: string, format: uuid }
                sessionId: { type: string, format: uuid }
                attemptId: { type: string, format: uuid }
                runtimeState: { enum: [duel_protocol_unavailable, ready] }
                connection: { type: "null" }
    MissionAttemptStatus:
      type: object
      required: [attemptId, sessionId, status, resultId, reconnect]
      properties:
        attemptId: { type: string, format: uuid }
        sessionId: { type: string, format: uuid }
        status: { enum: [queued, active, paused, completed, failed] }
        resultId: { type: [string, "null"], format: uuid }
        reconnect: { type: object, additionalProperties: true }
    ActiveGameplay:
      oneOf:
        - type: object
          additionalProperties: false
          required: [mode, attempt]
          properties:
            mode: { const: pve }
            attempt: { $ref: "#/components/schemas/MissionAttemptStatus" }
        - type: object
          additionalProperties: false
          required: [mode, matchmakingTicket, attempt]
          properties:
            mode: { const: pvp }
            matchmakingTicket: { $ref: "#/components/schemas/MatchmakingTicket" }
            attempt:
              oneOf:
                - { $ref: "#/components/schemas/MissionAttemptStatus" }
                - { type: "null" }
      discriminator:
        propertyName: mode
...
    LegacyBuildImport:
      type: object
      additionalProperties: false
      required: [schemaVersion, sourceBuildId, name, frameId, parts]
      properties:
        schemaVersion: { const: 3 }
        sourceBuildId: { type: string, maxLength: 128 }
        name: { type: string, minLength: 1, maxLength: 64 }
        frameId: { type: string, maxLength: 128 }
        cabin: { type: object, additionalProperties: true }
        parts:
          type: array
          maxItems: 256
          items:
            type: object
            required: [sourceInstanceId, kind, definitionId, gridX, gridY, rotation]
            properties:
              sourceInstanceId: { type: string }
              kind: { enum: [panel, module, element] }
              definitionId: { type: string }
              gridX: { type: integer, minimum: -64, maximum: 64 }
              gridY: { type: integer, minimum: -64, maximum: 64 }
              rotation: { enum: [0, 90, 180, 270] }